[ feature ] adiciona novos middlewares usuário não pode fazer transações para si mesmo não pode fazer transações com valores menor ou igual a zero deve haver saldo na conta a ser debitada para poder seguir com a transação

+++ backend/src/database/models/account.model.ts
@@ -1,3 +1,4 @@
+import { DECIMAL } from 'sequelize';
 import { Model, INTEGER } from 'sequelize';
 import db from '.';
 
@@ -15,7 +16,7 @@ Account.init({
   },
   balance: {
     allowNull: false,
-    type: INTEGER,
+    type: DECIMAL(10, 2),
   },
 }, {
   sequelize: db,
